turkreason

+++ open_models/validate.py
@@ -64,7 +64,7 @@ class TrainingConfig(BaseModel):
     push_to_private: bool = Field(True, description="Whether to push to private Hub")
 
     # Training hyperparameters
-    epochs: int = Field(1, description="Number of training epochs")
+    epochs: float = Field(1, description="Number of training epochs")
     max_steps: Optional[int] = Field(
         None, description="Maximum number of training steps"
     )
